Move LookupTableFields from common components to lookup-table components
