+++ graylog2-web-interface/src/components/event-definitions/event-definition-form/field-value-providers/LookupTableFieldValueProviderForm.jsx
@@ -18,7 +18,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import cloneDeep from 'lodash/cloneDeep';
 
-import { LookupTableFields } from 'components/common';
+import { LookupTableFields } from 'components/lookup-tables';
 
 class LookupTableFieldValueProviderForm extends React.Component {
   static propTypes = {
