+++ graylog2-web-interface/src/components/lookup-tables/LookupTableFields.tsx
